Improved the function that deletes previous test suites to handle 'file not found' errors more cleanly.  Improved print outs in all suite generators.

+++ scripts/_common.sh
@@ -30,9 +30,14 @@ function clearTestSuites() {
 	local language=$(echo $2 |tr [a-z] [A-Z]) #Programming language is fully capitalized
 	if [ -n "$3" ]; then local topic_type=$(tr '[:lower:]' '[:upper:]' <<< ${3:0:1})${3:1}${slash}; else local topic_type=""; fi #Topic type is capitalized 
 	echo "============================================================================================"
-	echo "Deleting: $(ls -1 $HOME/trunk/robotframework_SAL/$language/$topic_type${subsystem}_*)"
-    rm $HOME/trunk/robotframework_SAL/$language/$topic_type${subsystem}_*
-	echo "============================================================================================"
+	files=$(ls -1 $HOME/trunk/robotframework_SAL/$language/$topic_type${subsystem}_*)
+    if [ $? -eq 0 ]; then
+    	echo "Deleting:"
+		echo "$files"
+    	rm $HOME/trunk/robotframework_SAL/$language/$topic_type${subsystem}_*
+	else
+    	echo "Nothing to delete. Continuing."
+	fi
 	echo ""
 }
